Stateful firewall example added with the required hash functions

+++ src/hardware_dep/dpdk/includes/dpdkx_hash.h
@@ -5,4 +5,5 @@
 
 #include "dpdk_lib.h"
 
-void hash(uint16_t* result, enum_HashAlgorithm_t hash, uint16_t base, uint8_buffer_t data, uint32_t max, SHORT_STDPARAMS);
+//void hash(uint16_t* result, enum_HashAlgorithm_t hash, uint16_t base, uint8_buffer_t data, uint32_t max, SHORT_STDPARAMS);
+void hash(uint32_t* result, enum_HashAlgorithm_t hash, uint16_t base, uint8_buffer_t data, uint32_t max, SHORT_STDPARAMS);
